Update RenameTool.py
I used the tool and found it a bit confusing the replace and the replace with. 
So I changed it so that the current name could be changed to something they wanted. (but it's still in the code, if you want that)
And I added an underscore when you would add a pre- or suffix. Because I know in Maya that's the naming convention :)